requirements/dev.txt
-r base.txt
psycopg2-binary==2.9.7